osan 4 kirjautuminen toteutettu

+++ osa4/controllers/blogs.js
@@ -1,28 +1,81 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
-  response.json(blogs)
+  const blogs = await Blog
+    .find({})
+    .populate('user', { username: 1, name: 1 })
+
+  response.json(blogs.map(Blog.format))
 })
 
 blogsRouter.post('/', async (request, response) => {
-  const blog = new Blog(request.body)
-  if (blog.likes === undefined) {
-    blog.likes = 0
+  try {
+    const body = request.body
+
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if ( !request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    if (body.title === undefined || body.url === undefined) {
+      return response.status(400).json({ error: 'title or url missing' })
+    }
+
+    let likes = body.likes
+
+    if (likes === undefined) {
+      likes = 0
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = new Blog({
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes,
+      user: user._id
+    })
+
+    const savedBlog = await blog.save()
+
+    user.blogs = user.blogs.concat(savedBlog._id)
+
+    await user.save()
+
+    response.json(Blog.format(blog))
+  } catch (e) {
+    if (e.name === 'JsonWebTokenError') {
+      response.status(401).json({ error: e.message })
+    } else {
+      console.log(e)
+      response.status(500).json({ error: 'something went wrong' })
+    }
   }
-  if (blog.title === undefined || blog.url === undefined) {
-    return response.status(400).json(blog)
-  }
-  const result = await blog.save()
-  response.status(201).json(result)
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
   try {
-    await Blog.findByIdAndRemove(request.params.id)
+    const blog = await Blog.findById(request.params.id)
 
-    response.status(204).end()
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+    if ( !request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    if (blog.user.toString() === user._id) {
+      await Blog.findByIdAndRemove(request.params.id)
+      response.status(204).end()
+    } else {
+      response.status(400).json({ error: 'you can only delete your own blogs' })
+    }
   } catch (e) {
     console.log(e)
     response.status(400).send({ error: 'malformatted id' })
@@ -42,7 +95,7 @@ blogsRouter.put('/:id', async (request, response) => {
   try {
     await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
 
-    response.json(blog)
+    response.json(Blog.format(blog))
   } catch (e) {
     console.log(e)
     response.status(400).send({ error: 'malformatted id' })
